+++ mpitests/test_adjoint_jacobian.py
@@ -872,7 +872,7 @@ class TestAdjointJacobianQNode:
 
         jax = pytest.importorskip("jax")
         if dev.R_DTYPE == np.float64:
-            from jax.config import config  # pylint: disable=import-outside-toplevel
+            from jax import config  # pylint: disable=import-outside-toplevel
 
             config.update("jax_enable_x64", True)